Split misc utilities into ts and tsx
