+++ app/utils/cache.server.ts
@@ -14,7 +14,7 @@ import { remember } from '@epic-web/remember'
 import { LRUCache } from 'lru-cache'
 import { updatePrimaryCacheValue } from '#app/routes/resources/cache.sqlite.ts'
 import { getInstanceInfo, getInstanceInfoSync } from './litefs-js.server.js'
-import { getRequiredServerEnvVar } from './misc.tsx'
+import { getRequiredServerEnvVar } from './misc.ts'
 import { getUser } from './session.server.ts'
 import { time, type Timings } from './timing.server.ts'
