+++ app/utils/sitemap.server.ts
@@ -1,7 +1,7 @@
 import { type EntryContext } from 'react-router';
 import { type KCDHandle, type KCDSitemapEntry } from '#app/types.ts'
 import { isEqual } from '#app/utils/cjs/lodash.ts'
-import { getDomainUrl, removeTrailingSlash, typedBoolean } from './misc.tsx'
+import { getDomainUrl, removeTrailingSlash, typedBoolean } from './misc.ts'
 
 async function getSitemapXml(request: Request, remixContext: EntryContext) {
 	const domainUrl = getDomainUrl(request)
