+++ app/utils/testimonials.server.ts
@@ -4,7 +4,7 @@ import { pick } from '#app/utils/cjs/lodash.ts'
 import { cache, cachified } from './cache.server.ts'
 import { downloadFile } from './github.server.ts'
 import { markdownToHtml } from './markdown.server.ts'
-import { getErrorMessage, typedBoolean } from './misc.tsx'
+import { getErrorMessage, typedBoolean } from './misc.ts'
 import { type Timings } from './timing.server.ts'
 
 const allCategories = [
